ignore workplaces without sensor data
